requirements.txt
anyio==4.12.0
av==16.0.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
ctranslate2==4.6.1
faster-whisper==1.2.1
filelock==3.20.0
flatbuffers==25.9.23
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.41
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.6
humanfriendly==10.0
idna==3.11
mpmath==1.3.0
numpy==2.3.5
onnxruntime==1.23.2
packaging==25.0
protobuf==6.33.1
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
setuptools==78.1.1
shellingham==1.5.4
smmap==5.0.2
sympy==1.14.0
tokenizers==0.22.1
tqdm==4.67.1
typer-slim==0.20.0
typing_extensions==4.15.0
urllib3==2.5.0
